chore: fix initial setup (#3)
* build: update vscode settings and package.json
* build(npm): add and update dependencies

* ci(actions): run actions only on pr or push to master
* ci(actions): reorganize workflows

* chore: add lint action and improve npm scripts and vscode tasks
* chore(eslint): simplify configuration
* chore(vscode): change settings and tasks

* docs: write action metadata

* fix: reset files to begin development
