+++ .eslintrc.json
@@ -18,11 +18,17 @@
     "sourceType": "module"
   },
   "plugins": [
-    "jest",
-    "@typescript-eslint"
+    "@typescript-eslint",
+    "github",
+    "jest"
   ],
   "root": true,
   "rules": {
+    // ---------------------------------------------
+    // We rely on typescript's compiler noUnusedLocals option
+    "@typescript-eslint/no-unused-vars": "off",
+    "no-unused-vars": "off",
+    // ---------------------------------------------
     "@typescript-eslint/adjacent-overload-signatures": "error",
     "@typescript-eslint/array-type": "error",
     "@typescript-eslint/await-thenable": "error",
@@ -47,20 +53,7 @@
       "error",
       2
     ],
-    "@typescript-eslint/interface-name-prefix": "off",
-    "@typescript-eslint/member-delimiter-style": [
-      "error",
-      {
-        "multiline": {
-          "delimiter": "semi",
-          "requireLast": true
-        },
-        "singleline": {
-          "delimiter": "semi",
-          "requireLast": false
-        }
-      }
-    ],
+    "@typescript-eslint/member-delimiter-style": "error",
     "@typescript-eslint/no-array-constructor": "error",
     "@typescript-eslint/no-empty-function": "error",
     "@typescript-eslint/no-empty-interface": "error",
@@ -70,17 +63,15 @@
     "@typescript-eslint/no-inferrable-types": "error",
     "@typescript-eslint/no-misused-new": "error",
     "@typescript-eslint/no-namespace": "error",
-    "@typescript-eslint/no-non-null-assertion": "warn",
-    "@typescript-eslint/no-parameter-properties": "off",
+    "@typescript-eslint/no-non-null-assertion": "error",
     "@typescript-eslint/no-require-imports": "error",
     "@typescript-eslint/no-unnecessary-qualifier": "error",
     "@typescript-eslint/no-unnecessary-type-assertion": "error",
-    "@typescript-eslint/no-unused-vars": "error",
-    "@typescript-eslint/no-use-before-define": "off",
+    "@typescript-eslint/no-use-before-define": "error",
     "@typescript-eslint/no-useless-constructor": "error",
     "@typescript-eslint/no-var-requires": "error",
-    "@typescript-eslint/prefer-for-of": "warn",
-    "@typescript-eslint/prefer-function-type": "warn",
+    "@typescript-eslint/prefer-for-of": "error",
+    "@typescript-eslint/prefer-function-type": "error",
     "@typescript-eslint/prefer-includes": "error",
     "@typescript-eslint/prefer-namespace-keyword": "error",
     "@typescript-eslint/prefer-string-starts-ends-with": "error",
@@ -112,6 +103,10 @@
       "error",
       "smart"
     ],
+    "func-style": [
+      "error",
+      "expression"
+    ],
     "guard-for-in": "error",
     "id-blacklist": [
       "error",
@@ -140,12 +135,11 @@
     "no-bitwise": "error",
     "no-caller": "error",
     "no-cond-assign": "error",
-    "no-console": "off",
+    "no-console": "error",
     "no-debugger": "error",
     "no-empty": "error",
     "no-eval": "error",
-    "no-fallthrough": "off",
-    "no-invalid-this": "off",
+    "no-invalid-this": "error",
     "no-multiple-empty-lines": "error",
     "no-new-wrappers": "error",
     "no-shadow": [
@@ -170,7 +164,10 @@
     "prefer-const": "error",
     "radix": "error",
     "spaced-comment": "error",
-    "use-isnan": "error",
-    "valid-typeof": "off"
+    // TODO add and configure "import" plugin for advanced sorting
+    // ref: https://github.com/benmosher/eslint-plugin-import/blob/master/README.md
+    // ref: https://github.com/benmosher/eslint-plugin-import/blob/master/docs/rules/order.md
+    "sort-imports": "error",
+    "use-isnan": "error"
   }
 }
