+++ .github/workflows/check-files.yml
@@ -1,24 +1,28 @@
 name: Best practices for source files
 
-on: push
+on: # rebuild any PRs and main branch changes
+  pull_request:
+  push:
+    branches:
+      - master
 
 jobs:
   check-bom:
-    name: no BOMs found
+    name: No BOMs
     runs-on: ubuntu-latest
 
     steps:
       - uses: actions/checkout@v2
 
-      - name: check for BOMs
+      - name: Check for BOMs
         uses: arma-actions/bom-check@v1.0
 
   check-crlf:
-    name: no CRLF endings found
+    name: No CRLF endings
     runs-on: ubuntu-latest
 
     steps:
       - uses: actions/checkout@v2
 
-      - name: check for CRLF endings
+      - name: Check for CRLF endings
         uses: erclu/check-crlf@v1
