+++ .github/workflows/ci.yml
@@ -1,0 +1,48 @@
+name: Lint, Build and Test
+
+on: # rebuild any PRs and main branch changes
+  pull_request:
+  push:
+    branches:
+      - master
+
+jobs:
+  lint:
+    name: Run ESLint
+    runs-on: ubuntu-latest
+
+    steps:
+      - uses: actions/checkout@v2
+
+      - name: Run ESLint from package.json script
+        run: |
+          npm install
+          npm run lint --silent
+
+  # make sure build/ci work properly
+  build:
+    name: Build and test the code
+    runs-on: ubuntu-latest
+    needs: lint
+
+    steps:
+      - uses: actions/checkout@v2
+
+      - name: Run ESLint from package.json script
+        run: |
+          npm install
+          npm run all
+
+  # make sure the action works on a clean machine without building
+  test:
+    name: Test this action
+    runs-on: ubuntu-latest
+    needs: lint
+
+    steps:
+      - uses: actions/checkout@v2
+
+      - name: Run this action
+        uses: ./
+        with:
+          github-token: ${{ secrets.GITHUB_TOKEN }}
