+++ .github/workflows/conventional-commits.yml
@@ -1,17 +1,35 @@
 name: Conventional Commits
 
-on: push
+on: # rebuild any PRs and main branch changes
+  pull_request:
+    types:
+      - opened
+      - reopened
+      - synchronize # this triggers when pushing to pr branch
+      # - edited # this triggers e.g. when modifying pr title
+  push:
+    branches:
+      - master
 
 env:
   GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
 
 jobs:
+  # TODO find a way to check the pr title's conformity. these actions had some issues:
+  # amannn/action-semantic-pull-request@v1.1.1
+  # aslafy-z/conventional-pr-title-action@v2.2.0
   commitsar:
-    name: Check conformity to Conventional Commits specification
+    name: Check all commits
     runs-on: ubuntu-latest
+    # # do not run when the trigger has not changed any commit
+    # if: github.event.action != 'edited'
 
     steps:
       - uses: actions/checkout@v1
+      # This does not work
+      # - uses: actions/checkout@v2
+      #   with:
+      #     fetch-depth: 0
 
-      - name: run commitsar
+      - name: Run commitsar
         uses: outillage/commitsar@v0.9.0
