+++ .vscode/settings.json
@@ -1,0 +1,9 @@
+{
+  "files.exclude": {
+    "lib/**": true
+  },
+  "gulp.autoDetect": "off",
+  "jake.autoDetect": "off",
+  "grunt.autoDetect": "off",
+  "typescript.tsdk": "node_modules/typescript/lib",
+}
