+++ package.json
@@ -1,17 +1,18 @@
 {
   "name": "typescript-action",
   "version": "0.0.0",
-  "private": false,
+  "private": true,
   "description": "A file-based GitHub Action to assign collaborators to issues and pull requests",
   "main": "lib/main.js",
   "scripts": {
     "build": "tsc",
     "format": "prettier --write **/*.ts",
     "format-check": "prettier --check **/*.ts",
-    "lint": "tsc --noEmit && eslint src/**/*.ts --quiet --fix",
+    "lint": "eslint src/**/*.ts --quiet --max-warnings 0",
+    "lint-and-fix": "eslint src/**/*.ts --quiet --fix --max-warnings 0",
     "pack": "ncc build",
     "test": "jest",
-    "all": "npm run build && npm run format && npm run lint && npm run pack && npm test"
+    "all": "npm run build && npm run format && npm run lint-and-fix && npm run pack && npm test"
   },
   "repository": {
     "type": "git",
@@ -25,22 +26,28 @@
   "author": "Luca Ercole",
   "license": "MIT",
   "dependencies": {
-    "@actions/core": "^1.2.2"
+    "@actions/core": "^1.2.2",
+    "@actions/github": "^2.1.0",
+    "@actions/glob": "^0.1.0",
+    "@actions/io": "^1.0.2"
   },
   "devDependencies": {
-    "@types/jest": "^25.1.0",
-    "@types/node": "^13.5.0",
+    "@types/jest": "^25.1.1",
+    "@types/node": "^13.7.0",
     "@typescript-eslint/eslint-plugin": "^2.18.0",
     "@typescript-eslint/parser": "^2.18.0",
     "@zeit/ncc": "^0.21.0",
     "eslint": "^6.8.0",
     "eslint-plugin-github": "^3.4.0",
     "eslint-plugin-jest": "^23.6.0",
+    "flow-bin": "^0.117.0",
+    "graphql": "^14.6.0",
     "jest": "^25.1.0",
     "jest-circus": "^25.1.0",
     "js-yaml": "^3.13.1",
+    "nock": "^11.7.2",
     "prettier": "^1.19.1",
-    "ts-jest": "^25.0.0",
+    "ts-jest": "^25.1.0",
     "typescript": "^3.7.5"
   }
 }
