+++ src/main.ts
@@ -1,19 +1,9 @@
-import {getInput, debug, setOutput, setFailed} from "@actions/core";
-import {wait} from "./wait";
+import core from "@actions/core";
 
-async function run(): Promise<void> {
-  try {
-    const ms: string = getInput("milliseconds");
-    debug(`Waiting ${ms} milliseconds ...`);
+const run = async function(): Promise<void> {
+  await new Promise(() => null);
 
-    debug(new Date().toTimeString());
-    await wait(parseInt(ms, 10));
-    debug(new Date().toTimeString());
-
-    setOutput("time", new Date().toTimeString());
-  } catch (error) {
-    setFailed(error.message);
-  }
-}
+  core.setFailed("NOT IMPLEMENTED");
+};
 
 run();
