--- src/wait.ts
@@ -1,9 +1,0 @@
-export async function wait(milliseconds: number): Promise<string> {
-  return new Promise((resolve) => {
-    if (isNaN(milliseconds)) {
-      throw new Error("milliseconds not a number");
-    }
-
-    setTimeout(() => resolve("done!"), milliseconds);
-  });
-}
