+++ tsconfig.json
@@ -1,4 +1,5 @@
 {
+  // TODO should reduce this to bare minimum and use equivalent eslint rules
   "compilerOptions": {
     "allowUnreachableCode": false,
     "allowUnusedLabels": false,
@@ -19,6 +20,7 @@
     "target": "es6",
   },
   "exclude": [
+    "dist",
     "node_modules",
     "**/*.test.ts"
   ]
